requirements.txt
# CAIE搜题系统依赖包 - 简化版本

# PDF处理 (简化，避免编译问题)
PyPDF2==3.0.1

# OCR识别 (使用稳定版本)
paddleocr==2.6.1.3
opencv-python-headless==4.8.0.76
Pillow==9.5.0

# 数据处理
pandas==1.5.3
numpy==1.24.3

# Web框架
fastapi==0.100.1
uvicorn==0.23.2
python-multipart==0.0.6

# 搜索引擎
elasticsearch==8.8.0

# 缓存
redis==4.6.0

# AI模型 (轻量版)
sentence-transformers==2.2.2

# 工具库
requests==2.31.0
python-dotenv==1.0.0
huggingface_hub==0.10.1